Moved db.js to backend dir: Can now access modules
