+++ backend/controllers/gameResults.js
@@ -1,4 +1,4 @@
-const db = require("../../database/db");
+const db = require("../database/db");
 
 const getGameResults = async (req, res) => {
   try {
